Creation of modules by import/export + Babel7 Webpack4 implementation+ polyfill + fetch polyfill
