+++ .babelrc
@@ -1,3 +1,3 @@
-{
+ {
     "presets": ["@babel/preset-env"]
-}
+  }
