+++ client/src/index.html
@@ -31,14 +31,8 @@
 
                 <!-- Collect the nav links, forms, and other content for toggling -->
                 <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-                    <ul class="nav navbar-nav">
+                    <ul class="nav navbar-nav primary_navigation">
                         <li class="active"><a onclick="window.location.href=this">Home <span class="sr-only">(current)</span></a></li>
-                        <li><a>CNBC</a></li>
-                        <li><a>Time</a></li>
-                        <li><a>USA Today</a></li>
-                        <li><a>The new york times</a></li>
-                        <li><a>CNN</a></li>
-                        <li><a>Associated</a></li>
                     </ul>
                 </div>
                 <!-- /.navbar-collapse -->
