+++ client/src/styles/style.css
@@ -104,7 +104,7 @@ header>h1 {
 
 @media (max-width: 40em) {
     .thumbnail {
-        width: 100%;
+        width: 90%;
     }
     #newsArticlesList>li {
         display: block;
