+++ package.json
@@ -6,23 +6,26 @@
     "scripts": {
         "dev": "webpack --mode development",
         "build": "webpack --mode production",
-        "start": "webpack-dev-server --output-public-path=/dist/"
+        "start": "webpack && webpack-dev-server --output-public-path=/dist/"
     },
     "author": "",
     "license": "MIT",
     "devDependencies": {
-        "@babel/core": "^7.4.0",
-        "@babel/preset-env": "^7.4.2",
+        "@babel/core": "^7.4.3",
+        "@babel/polyfill": "^7.4.3",
+        "@babel/preset-env": "^7.4.3",
         "babel-loader": "^8.0.5",
         "babel-plugin-syntax-dynamic-import": "^6.18.0",
-        "babel-polyfill": "^6.26.0",
         "css-loader": "^2.1.1",
         "html-loader": "^0.5.5",
         "html-webpack-plugin": "^3.2.0",
         "style-loader": "^0.23.1",
-        "uglifyjs-webpack-plugin": "^2.1.2",
         "webpack": "^4.29.6",
         "webpack-cli": "^3.3.0",
-        "webpack-dev-server": "^3.2.1"
+        "webpack-dev-server": "^3.2.1",
+        "whatwg-fetch": "^3.0.0"
+    },
+    "dependencies": {
+        "core-js": "^3.0.0"
     }
 }
